+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `BulkSignature.list` endpoint
 - Fix `BulkSignature.scenarioVersion` type
 - Add `EnvelopeRecipientApproval` resource
+- Add `AccountIdentifyScenariosEndpoint` endpoint
 
 ## [2.6.0] - 2025-03-06
 - Edit `AccountGuide.settingsFilled`